feat: editor page redesign - theme, icons, toolbar
- Fix editor dark mode: WebView background now transparent, editor.html sets body bg
- Fix TagEditor crash: Color(tag.color.toInt()) instead of toULong()
- Replace weather emoji with Material Icons (WbSunny, Cloud, etc.)
- Redesign MetadataButton: bigger, with icons, better styling
- Redesign EditorToolbar: default expanded, bigger buttons, collapse icon
- Bump version to 1.7.0

+++ app/src/main/java/com/diary/app/ui/editor/TagEditor.kt
@@ -45,7 +45,7 @@ fun TagEditor(
     ) {
         allTags.forEach { tag ->
             val isSelected = tag.id in selectedTagIds
-            val tagColor = Color(tag.color.toULong())
+            val tagColor = Color(tag.color.toInt())
 
             Box(
                 modifier = Modifier
